refactor(models): remove any models files, and re-add some of the interfaces to the correct file location
