--- web/src/common/components/GlassRectangle/models.ts
@@ -1,12 +1,0 @@
-export interface IGlassContainer {
-  boxDarkShadowBlur: number;
-  boxDarkShadowHorizontalOffset: number;
-  boxDarkShadowVerticalOffset: number;
-  boxLightShadowBlur: number;
-  boxLightShadowHorizontalOffset: number;
-  boxLightShadowVerticalOffset: number;
-}
-
-export interface IGlassImageWrapper {
-  opacity?: number;
-}
