--- web/src/features/landing/components/WorkSection/models.ts
@@ -1,8 +1,0 @@
-export interface IWorkContainer {
-  workDescription: string;
-  workImageAlt: string;
-  workImageSrc: string;
-  workLinkContent: string;
-  workLinkPath: string;
-  workTitle: string;
-}
